% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_from_node_Detials.R
\name{gene_from_node_Detials}
\alias{gene_from_node_Detials}
\title{get gene names from Details property of chromosome}
\usage{
gene_from_node_Detials(details, re = "string")
}
\arguments{
\item{details}{the property in chromosome node}

\item{re}{default as 'string', means the result is a string, 'vector' means the result is a vector included all gene names.}
}
\value{
A string of genenames or a vector of genenames
}
\description{
Because Details property in chromosome have many information about genes. Then I make a function to get gene names from Details property of chromosome
}
\author{
Xiaowei
}
