% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_data.R
\name{geo_data}
\alias{geo_data}
\title{input GEO accession to get expression, phenotype and annotation data}
\usage{
geo_data(geo, geneID = "SYMBOL")
}
\arguments{
\item{geo}{a character, one geo accession}

\item{geneID}{a character, default as "SYMBOL". So you can choose one of "SYMBOL", "ENTREZ" and "RefSeq"}
}
\value{
a list which included expression data, phenotype data, annotation data and GPL number
}
\description{
input geo accession and gene id which you want, then this function will help you get all expression data which included gene id you want had converted.
}
\details{
if annotation not has your input gene id, then will be only return raw expression data. Also, I should remind you, after mapped gene id, I do not remove those
data not successful map the gene id, which is NA I do not remove.
}
\author{
xiaowei
}
