% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllGeneName.R
\name{getAllGeneName}
\alias{getAllGeneName}
\title{get all TF, LncRNA, genes in chromosome in GREG}
\usage{
getAllGeneName(node_labels)
}
\arguments{
\item{node_labels}{one of "genesInchr", "TF", "LncRNA", "all", "chr1", "chr2", "chr3" ... "chr22", "chrX" and "chrY"}
}
\value{
a data.frame which included geneName, NodeID, NodeLabel and NodeNames
}
\description{
get all TF, LncRNA, genes in chromosome in GREG
}
\examples{
library(RNeo4j)
graph = startGraph("http://localhost:7474/db/data/", username="neo4j", password="xiaowei")
AllgeneName_TF_LncRNA <- getAllGeneName('all')  #get all nodes that contains TF, lncRNA and gene names
TFName <- getAllGeneName('TF')                  #get all nodes of TF names
LncRNAName <- getAllGeneName('LncRNA')          #get all nodes of LncRNA names
AllgeneName <- getAllGeneName('genesInChr')     #get all nodes of gene names in chromosome
chr1 <- getAllGeneName('chr1')                  #get one chromosome nodes of genes names
}
\author{
Xiaowei
}
