% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPL_biocPackage.R
\name{GPL_biocPackage}
\alias{GPL_biocPackage}
\title{To know what bioconductor package is releted data from GEO database}
\usage{
GPL_biocPackage(GPL)
}
\arguments{
\item{GPL}{character, one GPL aceession}
}
\value{
a character or NA
}
\description{
You can easily input GPL aceession and return the annotation package names.
}
\examples{
GPL_biocPackage('GPL20')
GPL_biocPackage('GPL33')
}
