% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.download.R
\name{kegg.download}
\alias{kegg.download}
\title{Download pathway from kegg}
\usage{
kegg.download(pathway.list)
}
\arguments{
\item{pathway.list}{a character vector, details see: KEGGREST::keggList()}
}
\value{
A list wrapping a KEGG flat file.
}
\description{
Download pathway from kegg with input the organism. Retrieves all entries from the KEGG database for a set of KEGG identifers.
}
\details{
Because KEGGREST package only return 10 results at once. So I made it again to get results
ignored the numbers of query. So this functions is extended the function KEGGREST::keggGet()
}
\examples{
library(XWRpackage)
pathway.list <- keggList("pathway", organism = "hsa")
kegg.path <- kegg.download(pathway.list = pathway.list[1:11])






}
\author{
xiaowei
}
