% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.geneSet.R
\name{kegg.geneSet}
\alias{kegg.geneSet}
\title{download pathway of one organism from KEGG and make it as GeneSetCollection}
\usage{
kegg.geneSet(organism = "hsa", outputfile = NULL)
}
\arguments{
\item{organism}{A KEGG organism code from https://www.kegg.jp/kegg/catalog/org_list.html . for example, human is hsa, Mus musclus(mouse) is mmu.}

\item{outputfile}{strings for GMT format file. Default as NULL. If it is not NULL, it will save GMT file in your working directory.}
}
\value{
A GeneSetCollection object, GeneSetCollection class see GSEABase package. And GMT file.
}
\description{
download pathway of one organism from KEGG and make it as GeneSetCollection
}
\details{
We always use GeneSet/GeneSetCollection in GeneSet Enrichment Analysis, so it is important to make those pathway as GeneSet.
}
\examples{
library(XWRpackage)
pathway.list <- keggList("pathway", organism = "hsa")
kegg.path <- kegg.download(pathway.list = pathway.list[1:11])
kegg.geneSet <- path2geneSet(kegg.path[[1]])

hsa.geneSet <- kegg.geneSet("hsa") # Homo sapiens (human)
ptr.geneSet <- kegg.geneSet("ptr", outputfile = "ptr.gmt") # Pan troglodytes (chimpanzee)
ggo.geneSet <- kegg.geneSet("ggo") # Gorilla gorilla gorilla (western lowland gorilla)
pon.geneSet <- kegg.geneSet("pon") # Nomascus leucogenys (northern white-cheeked gibbon)



}
\author{
xiaowei
}
