% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nodes_by_genomic_region.R
\name{map_nodes_by_genomic_region}
\alias{map_nodes_by_genomic_region}
\title{input genomic region to get the node}
\usage{
map_nodes_by_genomic_region(g1)
}
\arguments{
\item{g1}{data.frame  likes g1 object which included nodellabel, start_number and end_number columns}
}
\value{
A data.frame  included all nodes' information about your input data
}
\description{
This function help you get the node about your input genomic region
}
\examples{
library(RNeo4j)
graph = startGraph("http://localhost:7474/db/data/", username="neo4j", password="xiaowei")
data("g123")
g1_nodes <- map_nodes_by_genomic_region(g1)
g2_nodes <- map_nodes_by_genomic_region(g2)
g3_nodes <- map_nodes_by_genomic_region(g3)
}
\author{
Xiaowei
}
