% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_rel_by_genomic_region.R
\name{map_rel_by_genomic_region}
\alias{map_rel_by_genomic_region}
\title{get all bind relationships with a same node via define a genomic regions, relationship and celltype}
\usage{
map_rel_by_genomic_region(g1, rel = "Interaction", celltype = "'IMR90'")
}
\arguments{
\item{g1}{A data.frame  likes g1 object which included nodellabel, start_number and end_number columns}

\item{rel}{default as "Interaction", you can choose "Interaction" or "Bind"}

\item{celltype}{You can choose celltype one of "'MCF7'", "'K562'", "'IMR90'", "'A549'", "'HELA'", "'H1ESC'", "'IPS6.9'" and "'IPS19.11'"}
}
\value{
results is a list about your input data, and each genename in your input data as a inside list which has two data.frame.
                  one is ADRelationship object,
                  the other is ADRelationship_summary object, which sum of ADRelationship object.
}
\description{
In this function, we can get interaction or bind relationships with a same nodes in different cell.
When you choose Interation relationship, the path is (A)-[r1:Inclusion]->(B)-[r2:Interaction]-(C)<-[r3:Inclusion]-(D)
In this path relationship, A and D nodes are chromosome nodes In GREG, B and C nodes are chromosome Range nodes.
When you choose Bind relationship, the path is (A)-[r:Bind]-(B)
In this path relationship, A nodes are chromosome nodes In GREG,but B nodes are TF or LncRNA nodes.
      B_label, B_Start, B_End, C_label,C_Start, C_End, D_label and D_Name
}
\examples{
library(RNeo4j)
graph = startGraph("http://localhost:7474/db/data/", username="neo4j", password="xiaowei")
data("g123")
g2_rel_dna_dna <- map_rel_by_genomic_region(g2)
g3_rel_dna_TF.LncRNA <- map_rel_by_genomic_region(g3, rel = "Bind")
}
\author{
Xiaowei
}
