% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_rel_of_both_genomic_region.R
\name{map_rel_of_both_genomic_region}
\alias{map_rel_of_both_genomic_region}
\title{get the relationship between genomic regions in the same input table.}
\usage{
map_rel_of_both_genomic_region(g1, rel = "Interaction", celltype = "'IMR90'")
}
\arguments{
\item{g1}{A data.frame  likes g1 object which included nodellabel, start_number and end_number columns}

\item{rel}{Only suppost "Interaction"}

\item{celltype}{You can choose celltype one of "'MCF7'", "'K562'", "'IMR90'", "'A549'", "'HELA'", "'H1ESC'", "'IPS6.9'" and "'IPS19.11'"}
}
\value{
A data.frame  which included all relationships information above-mentioned. colnames are A_label,A_Name, B_label, B_Start, B_End, C_label,
            C_Start, C_End, D_label and D_Name
}
\description{
get the relationship between genomic regions in the same input table.
}
\details{
The relationship is (A)-[r1:Inclusion]->(B)-[r2:Interaction]->(C)<-[r3:Inclusion]-(D)
So A and D nodes are chromosome nodes In GREG, B and C nodes are chromosome Range nodes.
}
\examples{
library(RNeo4j)
graph = startGraph("http://localhost:7474/db/data/", username="neo4j", password="xiaowei")
data("g123")
g1_rel <- map_rel_of_both_genomic_region(g1)
}
\author{
Xiaowei
}
