% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb.geneSet.R
\name{msigdb.geneSet}
\alias{msigdb.geneSet}
\title{download geneset of one species from MsigDB and make it as GeneSetCollection}
\usage{
msigdb.geneSet(
  species,
  category = NULL,
  subcategory = NULL,
  geneIdType = "entrez",
  outputfile = NULL
)
}
\arguments{
\item{species}{Species name, such as Homo sapiens or Mus musculus. Details see: msigdbr_species()}

\item{category}{MSigDB collection abbreviation, such as H or C1.}

\item{subcategory}{MSigDB sub-collection abbreviation, such as CGP or BP. Details see: msigdbr_collections()}

\item{geneIdType}{Default as "entrez". one of "entrez" and "symbol"}

\item{outputfile}{strings for GMT format file. Default as NULL. If it is not NULL, it will save GMT file in your working directory.}
}
\value{
A GeneSetCollection object, GeneSetCollection class see GSEABase package. And GMT file.
}
\description{
download geneset of one species from MsigDB and make it as GeneSetCollection
}
\details{
We always use GeneSet/GeneSetCollection in GeneSet Enrichment Analysis, so it is important to make those pathway as GeneSet.
}
\examples{
library(XWRpackage)

hsa.msigdb <- msigdb.geneSet(species = "Homo sapiens", category = "C2", subcategory = "CP:KEGG")
hsa.msigdb <- msigdb.geneSet(species = "Homo sapiens", category = "C2", subcategory = "CP:KEGG", outputfile = "msigdb.human.kegg.gmt")
hsa.msigdb <- msigdb.geneSet(species = "Homo sapiens", category = "C5", subcategory = ""GO:BP,GO:CC,GO:MF"", geneIdType = "symbol")


}
\author{
xiaowei
}
