% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path2geneSet.R
\name{path2geneSet}
\alias{path2geneSet}
\title{make the result of kegg.download to GeneSet}
\usage{
path2geneSet(kegg.path)
}
\arguments{
\item{kegg.path}{the result of kegg.download()}
}
\value{
A GeneSet object, GeneSet class see GSEABase package.
}
\description{
make the result of kegg.download to GeneSet
}
\details{
We always use GeneSet in GeneSet Enrichment Analysis, so it is important to make those pathway as GeneSet.
}
\examples{
library(XWRpackage)
library(KEGGREST)
pathway.list <- keggList("pathway", organism = "hsa")
kegg.path <- kegg.download(pathway.list = pathway.list[1:11])
kegg.geneSet <- path.to.geneSet(kegg.path[[1]])

}
\author{
xiaowei
}
